+++ doc/history.md
@@ -9,6 +9,7 @@ Artinya belum selesai, masih banyak bugs dan sedang proses development
 + Lib `Telegram` ditambahkan fowardMessage, sendMessage support entities dan reply
 + `main` dan `botapi` diperbaiki, digabungkan mengakses pada `./update`
 + memisahkan folder userbot dan botapi
++ dijadiin satu semua ada pada `main.js`
 
 ### v0.1.X